+++ src/deepboard/gui/components/__init__.py
@@ -2,7 +2,8 @@ from .chart_type import ChartType
 from .legend import Legend
 from .smoother import Smoother
 from .log_selector import LogSelector
-from .modal import Modal
+from .modal import Modal, WindowedModal
 from .split_selector import SplitSelector
 from .stat_line import StatLine
 from .artefact_group import ArtefactGroup
+from .quick_filter import QuickFilter