+++ CHANGELOG.md
@@ -8,6 +8,7 @@ data.
 - Hidden runs are now considered 'yanked' and are not taken into account when checking for duplicated runs. When you run an experiment and find out later there was a bug, you can hide it and re-run without worrying about duplicated runs.
 - Now can add and modify notes for each run.
 - Default theme for fragment tables
+- Quick filter: Now can filter categorical columns of the Result Table in the GUI
 
 ### Fixed bugs
 - Now can change the port and listening address of the GUI server